fixed : shiro bugs  #IDZ6T
